requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.1
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
Django==3.2.7
django-redis==5.0.0
djangorestframework==3.12.4
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
msgpack==1.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pyOpenSSL==21.0.0
pytz==2021.3
redis==3.5.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
zope.interface==5.4.0
